requirements.txt
python-docx>=0.8.11
openai>=1.0.0
pyyaml>=6.0
click>=8.0
tenacity>=8.0
pytest>=7.0
pytest-cov>=4.0
